complete

+++ start_all.sh
@@ -1,5 +1,26 @@
 #!/bin/bash
 
+
+# Проверка, что пользователь не администратор (root)
+if [ "$(id -u)" -eq 0 ]; then
+    echo "Ошибка: Этот скрипт нельзя запускать с правами администратора (root)." >&2
+    exit 1
+fi
+
+# Проверка операционной системы (должна быть Linux)
+if [ "$(uname -s)" != "Linux" ]; then
+    echo "Ошибка: Этот скрипт предназначен только для Linux." >&2
+    exit 1
+fi
+
+# Проверка, что интерпретатор - Bash (может быть вызван как sh, но должен быть bash)
+if [ -z "$BASH_VERSION" ]; then
+    echo "Ошибка: Этот скрипт должен запускаться в интерпретаторе Bash." >&2
+    exit 1
+fi
+
+echo '...Запуск всех систем...'
+
 SHOOTED_TARGETS_DIR="tmp/shooted_targets"
 DETECTED_TARGETS_DIR="tmp/detected_targets"
 rm -rf $SHOOTED_TARGETS_DIR 2>/dev/null
